refatoring alert box and using alert
