+++ packages/grafana-ui/src/components/Alert/Alert.tsx
@@ -1,21 +1,45 @@
 import React, { FC, ReactNode } from 'react';
+import classNames from 'classnames';
+import { AppNotificationSeverity } from 'app/types';
 
 interface Props {
   title: string;
   button?: {
-    text: string;
+    text?: string;
     onClick: (event: React.MouseEvent) => void;
   };
+  severity?: AppNotificationSeverity;
   children?: ReactNode;
+  icon?: string;
+}
+
+function getIconFromSeverity(severity: AppNotificationSeverity = AppNotificationSeverity.Error): string {
+  switch (severity) {
+    case AppNotificationSeverity.Error: {
+      return 'fa fa-exclamation-triangle';
+    }
+    case AppNotificationSeverity.Warning: {
+      return 'fa fa-exclamation-triangle';
+    }
+    case AppNotificationSeverity.Info: {
+      return 'fa fa-info-circle';
+    }
+    case AppNotificationSeverity.Success: {
+      return 'fa fa-check';
+    }
+    default:
+      return '';
+  }
 }
 
 export const Alert: FC<Props> = props => {
-  const { title, button, children } = props;
+  const { title, button, children, severity, icon } = props;
+  const alertClass = classNames('alert', `alert-${severity}`);
   return (
     <div className="alert-container">
-      <div className="alert-error alert">
+      <div className={alertClass}>
         <div className="alert-icon">
-          <i className="fa fa-exclamation-triangle" />
+          <i className={icon || getIconFromSeverity(severity)} />
         </div>
         <div className="alert-body">
           <div className="alert-title">{title}</div>
